Gradient Vector of the Multivariate Normal Distribution - Generic

Calculates gradient vector of the log of the likelihood function of the multivariate normal distribution for the ith observation.

grad_l_mvn_generic(x, mu, sigmacap)

Arguments

x
Numeric vector of length k. The ith vector of observations.

mu
Numeric vector. Parameter. Mean vector \(\boldsymbol{\mu}\).

sigmacap
Numeric matrix. Parameter. Covariance matrix \(\boldsymbol{\Sigma}\).

Value

A vector.

See also

Examples

n <- 5
mu <- c(0, 0)
sigmacap <- matrix(
  data = c(
    1, 0.5, 0.5, 1
  ),
  nrow = 2
)

xcap <- as.data.frame(
  t(
    rmvn_chol(
      n = n,
      mu = mu,
      sigmacap = sigmacap
    )
  )
)

lapply(
  X = xcap,
  FUN = grad_l_mvn_generic,
  mu = mu,
  sigmacap = sigmacap
)

#> $V1
#> [1]  0.5219108 -2.1512204 -0.5304712 -0.4560785  1.6472079
#> 
#> $V2
#> [1]  0.9303661 -0.6027681 -0.2338761  0.1058716 -0.4850020
#> 
#> $V3
#> [1]  2.09539462 -0.06073945  1.52867264  0.53939354 -0.66482203
#> 
#> $V4
#> [1] -1.9444885  0.6269982  1.2238511 -0.5525241 -0.4701033
#> 
#> $V5
#> [1]  1.0401683 -1.0554827 -0.1256916 -0.4312130 -0.1096448
#>