Hessian Matrix of the Multivariate Normal Distribution - Generic

Calculates hessian matrix of the log of the likelihood function of the multivariate normal distribution for the ith observation.

hess_l_mvn_generic(x, mu, sigmacap)

Arguments

x
Numeric vector of length k. The ith vector of observations.

mu
Numeric vector. Parameter. Mean vector \(\boldsymbol{\mu}\).

sigmacap
Numeric matrix. Parameter. Covariance matrix \(\boldsymbol{\Sigma}\).

Value

A matrix.

See also

Examples

n <- 5
mu <- c(0, 0)
sigmacap <- matrix(
  data = c(
    1, 0.5, 0.5, 1
  ),
  nrow = 2
)

xcap <- as.data.frame(
  t(
    rmvn_chol(
      n = n,
      mu = mu,
      sigmacap = sigmacap
    )
  )
)

lapply(
  X = xcap,
  FUN = hess_l_mvn_generic,
  mu = mu,
  sigmacap = sigmacap
)

#> $V1
#>            [,1]       [,2]         [,3]         [,4]       [,5]
#> [1,] -1.3333333  0.6666667 -1.350301868  0.881452832 -0.1031509
#> [2,]  0.6666667 -1.3333333  0.675150934 -1.453452817  0.2063019
#> [3,] -1.3503019  0.6751509 -0.478597463  0.003781665  0.1177585
#> [4,]  0.8814528 -1.4534528  0.003781665  0.613888137 -0.6640013
#> [5,] -0.1031509  0.2063019  0.117758533 -0.664001333  0.8569685
#> 
#> $V2
#>            [,1]       [,2]       [,3]       [,4]      [,5]
#> [1,] -1.3333333  0.6666667 -1.7057042   3.793624 -1.470386
#> [2,]  0.6666667 -1.3333333  0.8528521  -3.176090  2.940772
#> [3,] -1.7057042  0.8528521 -1.2931811   3.964211 -1.658810
#> [4,]  3.7936241 -3.1760902  3.9642115 -10.208018  6.116229
#> [5,] -1.4703860  2.9407720 -1.6588104   6.116229 -5.597216
#> 
#> $V3
#>            [,1]       [,2]       [,3]       [,4]       [,5]
#> [1,] -1.3333333  0.6666667 -0.5398709  0.6998493 -0.2149569
#> [2,]  0.6666667 -1.3333333  0.2699354 -0.7548278  0.4299139
#> [3,] -0.5398709  0.2699354  0.6702935 -0.6055177  0.1351855
#> [4,]  0.6998493 -0.7548278 -0.6055177  1.6909339 -0.6455057
#> [5,] -0.2149569  0.4299139  0.1351855 -0.6455057  0.7502694
#> 
#> $V4
#>            [,1]       [,2]       [,3]       [,4]       [,5]
#> [1,] -1.3333333  0.6666667 -2.4060227  1.6855930 -0.2412908
#> [2,]  0.6666667 -1.3333333  1.2030113 -2.6473135  0.4825817
#> [3,] -2.4060227  1.2030113 -3.4528199  2.1527923 -0.2131912
#> [4,]  1.6855930 -2.6473135  2.1527923 -3.1649772  0.0692698
#> [5,] -0.2412908  0.4825817 -0.2131912  0.0692698  0.7142251
#> 
#> $V5
#>            [,1]       [,2]       [,3]       [,4]       [,5]
#> [1,] -1.3333333  0.6666667  0.6722299 -1.9793429  0.8216140
#> [2,]  0.6666667 -1.3333333 -0.3361149  1.4938439 -1.6432280
#> [3,]  0.6722299 -0.3361149  0.5499691  0.1090412 -0.1920129
#> [4,] -1.9793429  1.4938439  0.1090412 -0.9703164  0.9521556
#> [5,]  0.8216140 -1.6432280 -0.1920129  0.9521556 -1.1362597
#>